+++ test/clang-tidy/nodecpp-naked-ptr-field-func.cpp
@@ -1,8 +1,8 @@
-// RUN: %check_clang_tidy %s nodecpp-naked-ptr-func %t
+// RUN: %check_clang_tidy %s nodecpp-naked-ptr-field-func %t
 
 // FIXME: Add something that triggers the check here.
 void f();
-// CHECK-MESSAGES: :[[@LINE-1]]:6: warning: function 'f' is insufficiently awesome [nodecpp-naked-ptr-func]
+// CHECK-MESSAGES: :[[@LINE-1]]:6: warning: function 'f' is insufficiently awesome [nodecpp-naked-ptr-field-func]
 
 // FIXME: Verify the applied fix.
 //   * Make the CHECK patterns specific enough and try to make verified lines